test-update/requirements.txt
flask
pdfminer.six
unidecode
